% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxido.R
\docType{package}
\name{roxidoExample-package}
\alias{roxidoExample}
\alias{roxidoExample-package}
\title{roxidoExample: A Short Phrase in Title Case}
\description{
One paragraph description of what the package does.
}
\author{
\strong{Maintainer}: @ROXIDO_GIVEN_NAME@ @ROXIDO_FAMILY_NAME@ \email{@ROXIDO_EMAIL@}

Other contributors:
\itemize{
  \item Authors of the dependency Rust crates (see inst/AUTHORS file for details) [contributor]
}

}
\keyword{internal}
